Se implemento la pantalla de inicio para visualizar los platillos. :)

+++ app/src/main/java/edu/utez/mx/restaurante/ui/screens/EditDishScreen.kt
@@ -9,7 +9,6 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
@@ -94,7 +93,7 @@ fun EditDishScreenPreview() {
         id = 1,
         nombre = "Platillo de ejemplo",
         descripcion = "Esta es una descripción de ejemplo.",
-        precio = 120.50,
+        precio = 120.50.toDouble(),
         image = R.drawable.enchiladas
     )
     EditDishScreen(dish = sampleDish)
